# Backend/app.py
from flask import (
    Flask,
    render_template,
    request,
    jsonify,
    session,
    send_from_directory,
)
from flask_swagger_ui import get_swaggerui_blueprint
import os
import re
import datetime
import jwt
import pymysql
from werkzeug.security import generate_password_hash, check_password_hash
from upload import upload_bp


app = Flask(__name__)

SWAGGER_URL = "/swagger"
API_URL = "/swagger.yaml"

swaggerui_blueprint = get_swaggerui_blueprint(
    SWAGGER_URL,
    API_URL,
    config={
        "app_name": "EVulnScanner API"
    }
)

app.register_blueprint(swaggerui_blueprint, url_prefix=SWAGGER_URL)
app.register_blueprint(upload_bp)

# ==========================================
# 정상 라우트 공통 보안 헤더 적용
# /vuln/security-headers 는 "헤더가 빠진 상태"를 시연하기 위한 의도적
# 취약 엔드포인트라서, 다른 라우트와 달리 이 헤더들을 붙이지 않는다.
# (다른 취약점들과 동일하게 정상/취약 라우트를 분리하는 설계에 맞춤)
# ==========================================
VULN_SECURITY_HEADERS_PATH = "/vuln/security-headers"

@app.after_request
def set_security_headers(response):
    if request.path == VULN_SECURITY_HEADERS_PATH:
        return response

    response.headers.setdefault("Content-Security-Policy", "default-src 'self'")
    response.headers.setdefault("X-Frame-Options", "DENY")
    response.headers.setdefault("X-Content-Type-Options", "nosniff")
    response.headers.setdefault("Strict-Transport-Security", "max-age=31536000; includeSubDomains")
    response.headers.setdefault("Referrer-Policy", "strict-origin-when-cross-origin")
    return response


# 세션 암호화 키
SECRET_KEY = os.getenv("SECRET_KEY")
if not SECRET_KEY:
    raise RuntimeError("SECRET_KEY environment variable is not set.")

app.secret_key = SECRET_KEY

# JWT 설정
JWT_SECRET = os.getenv("JWT_SECRET")
if not JWT_SECRET:
    raise RuntimeError("JWT_SECRET environment variable is not set.")
JWT_ALGORITHM = "HS256"
JWT_EXPIRE_MINUTES = 30

# Rate Limit 설정
MAX_LOGIN_ATTEMPTS = 5
LOCK_TIME = 30  # 초

login_attempts = {}

def get_db():
    conn = pymysql.connect(
        host=os.getenv("MYSQL_HOST", "localhost"),
        user=os.getenv("MYSQL_USER", "user"),
        password=os.getenv("MYSQL_PASSWORD", "1234"),
        database=os.getenv("MYSQL_DATABASE", "evulnscanner"),
        charset="utf8mb4"
    )
    return conn

@app.route("/health")
def health():
    try:
        conn = get_db()
        conn.close()
        return jsonify({"status": "ok"}), 200
    except Exception as e:
        return jsonify({"status": "error", "detail": str(e)}), 503

@app.route("/")
def home():

    return render_template("index.html")

@app.route("/swagger.yaml")
def swagger_yaml():
    return send_from_directory(app.root_path, "swagger.yaml")



# 회원가입
@app.route("/signup", methods=["GET", "POST"])
def signup():

    if request.method == "POST":

        username = request.form.get("username", "").strip()
        password = request.form.get("password", "")

        # 입력값 검증
        if not username or not password:
            return "아이디와 비밀번호를 입력하세요."

        # 아이디 길이 제한
        if len(username) < 4 or len(username) > 20:
            return "아이디는 4~20자만 가능합니다."

        # 비밀번호 길이 제한햣
        if len(password) < 8 or len(password) > 20:
            return "비밀번호는 8~20자만 가능합니다."

        # 아이디 특수문자 제한
        if not re.fullmatch(r"[A-Za-z0-9_]+", username):
            return "아이디는 영문, 숫자, _(언더바)만 사용할 수 있습니다."

        # 비밀번호 해시
        hashed_password = generate_password_hash(password)

        conn = get_db()
        cursor = conn.cursor()

        cursor.execute(
            """
            INSERT INTO users(username,password)
            VALUES(%s,%s)
            """,
            (username, hashed_password)
        )

        conn.commit()
        conn.close()

        return "회원가입 성공!"

    return render_template("signup.html")




# 사용자 조회 (인증 테스트)
@app.route("/users")
def users():

    if "username" not in session:
        return jsonify({
            "error": "로그인이 필요합니다."
        }), 401

    conn = get_db()
    cursor = conn.cursor()

    cursor.execute(
        "SELECT id, username FROM users"
    )

    users = cursor.fetchall()

    conn.close()

    result = []

    for user in users:
        result.append({
            "id": user[0],
            "username": user[1]
        })

    return jsonify(result)





# 로그인
@app.route("/login", methods=["GET", "POST"])
def login():

    if request.method == "POST":

        username = request.form.get("username", "").strip()
        password = request.form.get("password", "")

        # 입력값 검증
        if not username or not password:
            return "아이디와 비밀번호를 모두 입력하세요."

    
        now = datetime.datetime.now()

        if username in login_attempts:
            attempt = login_attempts[username]

            # 잠금 시간이 지나지 않았으면 로그인 차단
            if (
                attempt["count"] >= MAX_LOGIN_ATTEMPTS
                and (now - attempt["last_attempt"]).seconds < LOCK_TIME
            ):
                remain = LOCK_TIME - (now - attempt["last_attempt"]).seconds
                return f"로그인 시도 횟수를 초과했습니다. {remain}초 후 다시 시도하세요."

            # 잠금 시간이 지나면 초기화
            if (now - attempt["last_attempt"]).seconds >= LOCK_TIME:
                login_attempts.pop(username)

        conn = get_db()
        cursor = conn.cursor()

        cursor.execute(
            """
            SELECT * FROM users
            WHERE username=%s
            """,
            (username,)
        )

        user = cursor.fetchone()

        conn.close()

        # 로그인 성공
        if user and check_password_hash(user[2], password):

            # 사용자 정보 세션 저장
            session["user_id"] = user[0]
            session["username"] = user[1]

            # 로그인 성공 시 실패 횟수 초기화
            login_attempts.pop(username, None)

            return "로그인 성공!"

        # 로그인 실패
        if username not in login_attempts:
            login_attempts[username] = {
                "count": 1,
                "last_attempt": now
            }
        else:
            login_attempts[username]["count"] += 1
            login_attempts[username]["last_attempt"] = now

        return "아이디 또는 비밀번호가 올바르지 않습니다."

    return render_template("login.html")

# ==========================================
# 의도적 취약점 예제: Broken Authentication
# 비밀번호 검증 없이 사용자 존재 여부만 확인
# 실제 서비스에서는 사용하면 안 됨
# ==========================================

@app.route("/vuln/broken-auth", methods=["POST"])
def vuln_broken_auth():

    username = request.form.get("username", "").strip()
    password = request.form.get("password", "")

    if not username or not password:
        return jsonify({
            "success": False,
            "message": "아이디와 비밀번호를 입력하세요."
        }), 400

    conn = get_db()
    cursor = conn.cursor()

    cursor.execute(
        """
        SELECT * FROM users
        WHERE username=%s
        """,
        (username,)
    )

    user = cursor.fetchone()

    conn.close()

    # -----------------------------
    # 의도적 취약점 
    # 비밀번호를 검증하지 않고
    # 사용자 존재 여부만 확인
    # -----------------------------
    if user:

        session["username"] = username

        return jsonify({
            "success": True,
            "message": "로그인 성공 (Broken Authentication)"
        })

    return jsonify({
        "success": False,
        "message": "존재하지 않는 사용자입니다."
    }), 404

# ==========================================
# 의도적 취약점 예제: Rate Limit Missing
# 로그인 요청 횟수를 제한하지 않음
# 실제 서비스에서는 사용하면 안 됨
# ==========================================

@app.route("/vuln/rate-limit", methods=["POST"])
def vuln_rate_limit():

    username = request.form.get("username", "").strip()
    password = request.form.get("password", "")

    if not username or not password:
        return jsonify({
            "success": False,
            "message": "아이디와 비밀번호를 입력하세요."
        }), 400

    conn = get_db()
    cursor = conn.cursor()

    cursor.execute(
        """
        SELECT * FROM users
        WHERE username=%s
        """,
        (username,)
    )

    user = cursor.fetchone()

    conn.close()

    # ----------------------------------
    # 의도적 취약점
    # 로그인 실패 횟수 제한 없음
    # 요청 횟수 제한 없음
    # ----------------------------------
    if user and check_password_hash(user[2], password):
        return jsonify({
            "success": True,
            "message": "로그인 성공"
        })

    return jsonify({
        "success": False,
        "message": "아이디 또는 비밀번호가 올바르지 않습니다."
    }), 401


# 로그아웃
@app.route("/logout")
def logout():

    session.clear()

    return """
    <script>
    alert("로그아웃되었습니다.");
    location.href="/";
    </script>
    """

# ==========================================
# JWT 발급 API
# 로그인 성공 시 JWT 발급
# ==========================================

@app.route("/api/token", methods=["POST"])
def issue_token():

    username = request.form.get("username", "").strip()
    password = request.form.get("password", "")

    if not username or not password:
        return jsonify({
            "success": False,
            "message": "아이디와 비밀번호를 입력하세요."
        }), 400

    conn = get_db()
    cursor = conn.cursor()

    cursor.execute(
        """
        SELECT * FROM users
        WHERE username=%s
        """,
        (username,)
    )

    user = cursor.fetchone()

    conn.close()

    if not user or not check_password_hash(user[2], password):
        return jsonify({
            "success": False,
            "message": "아이디 또는 비밀번호가 올바르지 않습니다."
        }), 401

    payload = {
        "username": username,
        "exp": datetime.datetime.utcnow() + datetime.timedelta(minutes=JWT_EXPIRE_MINUTES)
    }

    token = jwt.encode(
        payload,
        JWT_SECRET,
        algorithm=JWT_ALGORITHM
    )

    return jsonify({
        "success": True,
        "token": token
    })

# ==========================================
# 의도적 취약점 예제: JWT Validation Missing
# Authorization 헤더는 확인하지만
# JWT 서명 검증을 하지 않음
# ==========================================

@app.route("/vuln/profile")
def vuln_profile():

    auth_header = request.headers.get("Authorization")

    if not auth_header:
        return jsonify({
            "success": False,
            "message": "Authorization 헤더가 없습니다."
        }), 401

    # ----------------------------------
    # 의도적 취약점
    # JWT를 decode하거나 서명을 검증하지 않음
    # ----------------------------------

    token = auth_header.replace("Bearer ", "")

    return jsonify({
        "success": True,
        "message": "JWT 검증 없이 접근 성공",
        "received_token": token
    })


@app.route("/posts", methods=["GET","POST"])
def posts():


    conn = get_db()
    cursor = conn.cursor()



    if request.method == "POST":


        if "username" not in session:
            conn.close()
            return "로그인이 필요합니다."

        title = request.form.get("title", "").strip()
        content = request.form.get("content", "").strip()

        # 작성자는 로그인한 사용자로 고정
        writer = session["username"]
        user_id = session["user_id"]


        # 입력값 검증
        if not title or not content or not writer:

            conn.close()
            return "모든 값을 입력하세요."


        # 제목 길이 제한
        if len(title) > 100:

            conn.close()
            return "제목은 100자 이하만 가능합니다."


        cursor.execute(

            """
            INSERT INTO posts(title, content, writer, user_id)
            VALUES(%s, %s, %s, %s)
            """,

            (
                title,
                content,
                writer,
                user_id
            )

        )

        conn.commit()



    keyword = request.args.get("keyword","").strip()



    if keyword:


        cursor.execute(

            """
            SELECT * FROM posts
            WHERE title LIKE %s
            ORDER BY id DESC
            """,

            (f"%{keyword}%",)

        )


    else:


        cursor.execute(

            """
            SELECT * FROM posts
            ORDER BY id DESC
            """

        )



    posts = cursor.fetchall()


    conn.close()



    return render_template(

        "posts.html",

        posts=posts,

        keyword=keyword,

        current_user_id=session.get("user_id")

    )
# ==========================================
# 의도적 취약점 예제: Stored XSS
# 게시글 내용을 escaping 없이 출력
# ==========================================

@app.route("/vuln/posts")
def vuln_posts():

    conn = get_db()
    cursor = conn.cursor()

    cursor.execute(
        """
        SELECT * FROM posts
        ORDER BY id DESC
        """
    )

    posts = cursor.fetchall()

    conn.close()

    return render_template(
        "vuln_posts.html",
        posts=posts,
        current_user_id=session.get("user_id")
    )






# 게시글 API 전체 조회 + 검색
@app.route("/api/posts")
def api_posts():

    keyword = request.args.get("keyword", "")

    conn = get_db()
    cursor = conn.cursor()


    if keyword:

        cursor.execute(
            """
            SELECT * FROM posts
            WHERE title LIKE %s
            ORDER BY id DESC
            """,
            (f"%{keyword}%",)
        )

    else:

        cursor.execute(
            """
            SELECT * FROM posts
            ORDER BY id DESC
            """
        )


    posts = cursor.fetchall()


    conn.close()


    result = []


    for post in posts:

        result.append({

            "id": post[0],
            "title": post[1],
            "content": post[2],
            "writer": post[3]

        })


    return jsonify(result)






# 게시글 API 단일 조회
@app.route("/api/posts/<int:post_id>")
def api_post(post_id):


    conn=get_db()
    cursor=conn.cursor()


    cursor.execute(

        """
        SELECT * FROM posts
        WHERE id=%s
        """,

        (post_id,)

    )


    post=cursor.fetchone()


    conn.close()



    if post is None:


        return jsonify(
            {
                "error":"게시글 없음"
            }
        ),404



    return jsonify({

        "id":post[0],
        "title":post[1],
        "content":post[2],
        "writer":post[3]

    })

# 게시글 수정
@app.route("/posts/edit/<int:post_id>", methods=["GET", "POST"])
def edit_post(post_id):

    conn = get_db()
    cursor = conn.cursor()

    # 게시글 조회
    cursor.execute(
        """
        SELECT id, title, content, writer, user_id
        FROM posts
        WHERE id=%s
        """,
        (post_id,)
    )

    post = cursor.fetchone()

    if post is None:
        conn.close()
        return "게시글이 존재하지 않습니다."

    # 로그인 확인
    if "username" not in session:
        conn.close()
        return "로그인이 필요합니다."

    # 작성자 확인
    # 자신의 글만 수정 가능
    if post[4] != session["user_id"]:
        conn.close()
        return "작성자만 수정할 수 있습니다."

    if request.method == "POST":

        title = request.form.get("title", "").strip()
        content = request.form.get("content", "").strip()

        # 제목 검증
        if not title:
            conn.close()
            return "제목을 입력하세요."

        if len(title) > 100:
            conn.close()
            return "제목은 100자 이하만 가능합니다."

        # 내용 검증
        if not content:
            conn.close()
            return "내용을 입력하세요."

        cursor.execute(
            """
            UPDATE posts
            SET title=%s,
                content=%s
            WHERE id=%s
            """,
            (title, content, post_id)
        )

        conn.commit()
        conn.close()

        return """
        <script>
        alert("수정 완료");
        location.href="/posts";
        </script>
        """

    conn.close()

    return render_template(
        "edit_post.html",
        post=post
    )

# 게시글 삭제
@app.route("/posts/delete/<int:post_id>")
def delete_post(post_id):

    conn = get_db()
    cursor = conn.cursor()

    # 게시글 조회
    cursor.execute(
        """
        SELECT id, title, content, writer, user_id
        FROM posts
        WHERE id=%s
        """,
        (post_id,)
    )

    post = cursor.fetchone()

    if post is None:
        conn.close()
        return "게시글이 존재하지 않습니다."

    # 로그인 확인
    if "username" not in session:
        conn.close()
        return "로그인이 필요합니다."

    # 작성자 확인
    if post[4] != session["user_id"]:
        conn.close()
        return "작성자만 삭제할 수 있습니다."

    # 삭제
    cursor.execute(
        """
        DELETE FROM posts
        WHERE id=%s
        """,
        (post_id,)
    )

    conn.commit()
    conn.close()

    return """
    <script>
    alert("삭제 완료");
    location.href="/posts";
    </script>
    """



# ==========================================
# 의도적 취약점 예제: IDOR (게시글 수정)
# 권한 확인 없이 게시글 수정 가능
# ==========================================

@app.route("/vuln/posts/edit/<int:post_id>", methods=["GET", "POST"])
def vuln_edit_post(post_id):

    conn = get_db()
    cursor = conn.cursor()

    if request.method == "POST":

        title = request.form.get("title", "").strip()
        content = request.form.get("content", "").strip()

        cursor.execute(
            """
            UPDATE posts
            SET title=%s, content=%s
            WHERE id=%s
            """,
            (title, content, post_id)
        )

        conn.commit()
        conn.close()

        return """
        <script>
        alert("취약 수정 완료");
        location.href="/posts";
        </script>
        """

    cursor.execute(
        """
        SELECT * FROM posts
        WHERE id=%s
        """,
        (post_id,)
    )

    post = cursor.fetchone()

    conn.close()

    if post is None:
        return "게시글이 존재하지 않습니다."

    return render_template(
        "edit_post.html",
        post=post
    )


# ==========================================
# 의도적 취약점 예제: IDOR (게시글 삭제)
# 권한 확인 없이 게시글 삭제 가능
# ==========================================

@app.route("/vuln/posts/delete/<int:post_id>")
def vuln_delete_post(post_id):

    conn = get_db()
    cursor = conn.cursor()

    cursor.execute(
        """
        DELETE FROM posts
        WHERE id=%s
        """,
        (post_id,)
    )

    conn.commit()
    conn.close()

    return """
    <script>
    alert("취약 삭제 완료");
    location.href="/posts";
    </script>
    """



# ==========================================
# 의도적 취약점 예제: SQL Injection
# 사용자 입력값(username, password)을 검증 없이
# SQL 쿼리에 직접 삽입하는 취약한 코드
# 실제 서비스에서는 사용하면 안 됨
# ==========================================


@app.route("/vuln/login", methods=["POST"])
def vuln_login():


    username=request.form.get("username")
    password=request.form.get("password")


    conn=get_db()
    cursor=conn.cursor()


# 의도적 취약점 발생 부분
# Prepared Statement를 사용하지 않고
# 문자열 조합으로 SQL Query 생성
    query=f"""
    SELECT * FROM users
    WHERE username='{username}'
    AND password='{password}'
    """


    print(query)


    cursor.execute(query)


    user=cursor.fetchone()


    conn.close()



    if user:

        return jsonify({

            "success":True,
            "message":"로그인 성공"

        })


    return jsonify({

        "success":False,
        "message":"로그인 실패"

    })






# ==========================================
# 의도적 취약점 예제: Reflected XSS
# 사용자 입력값을 HTML escaping 없이 출력하는 취약한 코드
# 실제 서비스에서는 입력값 검증 및 escaping 필요
# ==========================================
@app.route("/vuln/comment")
def comment():


    text=request.args.get("text","")


    return f"""

    <h2>댓글</h2>

    <p>{text}</p>

    """

# ==========================================
# 의도적 취약점 예제: Security Headers 누락
# set_security_headers()가 이 경로만 예외 처리해서 CSP, X-Frame-Options,
# X-Content-Type-Options, Strict-Transport-Security, Referrer-Policy가
# 전부 빠진 응답을 그대로 돌려준다.
# 실제 서비스에서는 사용하면 안 됨
# ==========================================

@app.route("/vuln/security-headers")
def vuln_security_headers():

    return """

    <h2>보안 헤더 누락 페이지</h2>

    <p>이 페이지는 권장 보안 응답 헤더(CSP, X-Frame-Options, X-Content-Type-Options,
    Strict-Transport-Security, Referrer-Policy)가 의도적으로 빠져 있습니다.</p>

    """

# ==========================================
# 의도적 취약점 예제: Directory Traversal
# 사용자 입력을 파일 경로 검증 없이 사용
# 실제 서비스에서는 사용하면 안 됨
# ==========================================

@app.route("/vuln/download")
def vuln_download():

    filename = request.args.get("file")

    if not filename:
        return "파일명을 입력하세요."

    try:

        with open(filename, "r", encoding="utf-8") as f:
            return f.read()

    except Exception as e:
        return str(e)




if __name__ == "__main__":
    app.run(host="0.0.0.0", port=5000)

    