# Backend/swagger.yaml
openapi: 3.0.3

info:
  title: EVulnScanner API
  version: 1.0.0
  description: |
    EVulnScanner는 Flask 기반 웹 취약점 테스트용 프로젝트입니다.

    구현 기능
    - 회원가입
    - 로그인
    - 게시판 CRUD
    - 게시글 검색
    - 파일 업로드
    - 입력값 검증
    - JWT 발급 및 인증 테스트
    - 취약점 테스트용 API

servers:
  - url: http://localhost:5000

paths:

  /signup:
    get:
      summary: 회원가입 페이지
      responses:
        '200':
          description: 회원가입 페이지

    post:
      summary: 회원가입
      description: |
        아이디는 4~20자의 영문, 숫자, _(언더바)만 허용됩니다.
        비밀번호는 8~20자여야 합니다.

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string

      responses:
        '200':
          description: 회원가입 성공

  /login:
    get:
      summary: 로그인 페이지
      responses:
        '200':
          description: 로그인 페이지

    post:
      summary: 로그인
      description: |
        로그인 실패 횟수를 제한하는 Rate Limit이 적용되어 있습니다.

      requestBody:
        required: true

        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string

      responses:
        '200':
          description: 로그인 성공

        '429':
          description: 로그인 시도 횟수 초과
        
        '401':
          description: 아이디 또는 비밀번호 오류

  /users:
    get:
      summary: 사용자 목록 조회

      responses:
        '200':
          description: 사용자 목록

        '401':
          description: 로그인 필요

  /posts:
    get:
      summary: 게시글 조회 및 검색

      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string

      responses:
        '200':
          description: 게시글 목록

    post:
      summary: 게시글 작성

      requestBody:
        required: true

        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                writer:
                  type: string

      responses:
        '200':
          description: 작성 성공

  /posts/edit/{post_id}:
    get:
      summary: 게시글 수정 페이지

      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: 수정 페이지

    post:
      summary: 게시글 수정

      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string

      responses:
        '200':
          description: 수정 완료

  /posts/delete/{post_id}:
    get:
      summary: 게시글 삭제

      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: 삭제 완료

  /api/posts:
    get:
      summary: 게시글 API 전체 조회

      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string

      responses:
        '200':
          description: 게시글 목록(JSON)

  /api/posts/{post_id}:
    get:
      summary: 게시글 단일 조회

      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: 게시글 정보

        '404':
          description: 게시글 없음

  /upload:
    get:
      summary: 파일 업로드 페이지

      responses:
        '200':
          description: 업로드 페이지

    post:
      summary: 정상 파일 업로드

      requestBody:
        required: true

        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary

      responses:
        '200':
          description: 업로드 성공

  /vuln/login:
    post:
      summary: SQL Injection 취약 로그인
      description: SQL Injection 취약점 테스트용 API

      requestBody:
        required: true

        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string

      responses:
        '200':
          description: 로그인 결과

          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true

  /vuln/broken-auth:
    post:
      summary: Broken Authentication
      description: 비밀번호 검증을 수행하지 않는 취약 로그인 API

      requestBody:
        required: true

        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string

      responses:
        '200':
          description: 로그인 결과

  /api/token:
    post:
      summary: JWT 발급
      description: 로그인 성공 시 JWT를 발급합니다.

      requestBody:
        required: true

        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string

      responses:
        '200':
          description: JWT 발급 성공

          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIs...

        '401':
          description: 로그인 실패

  /vuln/profile:
    get:
      summary: JWT Validation Missing

      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer <JWT_TOKEN>
          schema:
            type: string

      responses:
        '200':
          description: 접근 성공

        '401':
          description: Authorization 헤더 없음

  /vuln/rate-limit:
    post:
      summary: Rate Limit Missing
      description: 로그인 요청 횟수를 제한하지 않는 취약 로그인 API

      requestBody:
        required: true

        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string

      responses:
        '200':
          description: 로그인 성공

        '401':
          description: 로그인 실패

  /vuln/comment:
    get:
      summary: Reflected XSS 테스트

      parameters:
        - name: text
          in: query
          schema:
            type: string

      responses:
        '200':
          description: 취약 페이지

  /vuln/posts:
    get:
      summary: Stored XSS 테스트

      responses:
        '200':
          description: Stored XSS 취약 게시판

  /vuln/posts/edit/{post_id}:
    get:
      summary: IDOR 취약 수정 페이지
      description: 권한 검사 없이 게시글 수정 페이지를 조회

      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: 수정 페이지

    post:
      summary: IDOR 취약 수정

      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
            
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string

      responses:
        '200':
          description: 수정 완료

  /vuln/posts/delete/{post_id}:
    get:
      summary: IDOR 취약 삭제

      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: 삭제 완료

  /vuln/upload:
    post:
      summary: 취약 파일 업로드
      description: 확장자 검증 없이 업로드

      requestBody:
        required: true

        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary

      responses:
        '200':
          description: 업로드 성공

  /vuln/download:
    get:
      summary: Directory Traversal 테스트
      description: |
        file 파라미터를 통해 파일을 다운로드합니다.
        ../ 등을 이용한 Directory Traversal 테스트용 API입니다.

      parameters:
        - name: file
          in: query
          required: true
          schema:
            type: string

      responses:
        '200':
          description: 파일 내용 반환

  /vuln/security-headers:
    get:
      summary: Security Headers 누락 테스트
      description: |
        CSP, X-Frame-Options, X-Content-Type-Options, Strict-Transport-Security,
        Referrer-Policy 등 권장 보안 응답 헤더가 의도적으로 빠져 있는 페이지입니다.
        다른 정상 라우트에는 이 헤더들이 공통으로 적용되어 있습니다.

      responses:
        '200':
          description: 보안 헤더가 빠진 취약 페이지